ci: add mypy + refactor common testing
