+++ .pre-commit-config.yaml
@@ -9,3 +9,7 @@ repos:
     rev: v0.0.275
     hooks:
       - id: ruff
+  - repo: https://github.com/pre-commit/mirrors-mypy
+    rev: 'v1.4.1'
+    hooks:
+      - id: mypy
