+++ pyproject.toml
@@ -12,4 +12,8 @@ dependencies = [
   "pytest",
   "toml",
   "pre-commit",
+  "mypy",
 ]
+
+[tool.ruff.pydocstyle]
+convention = "numpy"
